my_WAGTAIL_site/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.6.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-debug-toolbar==2.2
django-extensions==2.2.9
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.9
l18n==2018.5
Pillow==6.2.2
pytils==0.3
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.8.2
webencodings==0.5.1
Willow==1.3
